requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
sqlparse==0.5.5
django-allauth==65.14.0
django-oauth-toolkit==3.2.0
google-auth==2.47.0
requests==2.32.5
PyJWT==2.10.1
